+++ src/heavyedge/plugins/profile.py
@@ -163,49 +163,6 @@ class OutlierCommand(Command):
         self.logger.info(f"Removed outliers: {out.path}")
 
 
-@register_command("mean", "Compute mean profile")
-class MeanCommand(Command):
-    def add_parser(self, main_parser):
-        mean = main_parser.add_parser(
-            self.name,
-            description="Compute mean profile and save as hdf5 file.",
-            epilog="The resulting hdf5 file is in 'ProfileData' structure.",
-        )
-        mean.add_argument(
-            "profiles",
-            type=pathlib.Path,
-            help="Path to preprocessed profile data in 'ProfileData' structure.",
-        )
-        mean.add_config_argument(
-            "--wnum",
-            type=int,
-            help="Number of sample points to compute Wasserstein mean.",
-        )
-        mean.add_argument("-o", "--output", type=pathlib.Path, help="Output file path")
-
-    def run(self, args):
-        import numpy as np
-
-        from heavyedge.api import mean
-        from heavyedge.io import ProfileData
-
-        self.logger.info(f"Averaging: {args.profiles}")
-
-        with ProfileData(args.profiles) as data:
-            _, M = data.shape()
-            x = data.x()
-            res = data.resolution()
-            name = data.name()
-            pmean = mean(x, data.profiles(), args.wnum)
-
-        with ProfileData(args.output, "w").create(M, res, name) as out:
-            L = len(pmean)
-            pmean = np.pad(pmean, (0, M - L), constant_values=np.nan)
-            out.write_profiles(pmean.reshape(1, -1), [L], [name])
-
-        self.logger.info(f"Averaged: {out.path}")
-
-
 @register_command("merge", "Merge profile data")
 class MergeCommand(Command):
     def add_parser(self, main_parser):